Add files via upload
Shrank the height of the pattern 0.8mm on each side.  I think the vinyl I was using on my first cut stretched a little bit as I applied it, so if using the same vinyl, it's better if it's a bit smaller.  Added my first attempt at the windows logo cutout and personalization with name.  I don't have the vinyl I ordered yet, so this pattern is untested.